+++ co-log/src/Colog/Actions.hs
@@ -1,12 +1,7 @@
 module Colog.Actions
-       ( -- * 'String' actions
-         logStringStdout
-       , logStringStderr
-       , logStringHandle
-       , withLogStringFile
-
+       (
          -- * 'ByteString' actions
-       , logByteStringStdout
+         logByteStringStdout
        , logByteStringStderr
        , logByteStringHandle
        , withLogByteStringFile
@@ -18,38 +13,10 @@ module Colog.Actions
        , withLogTextFile
        ) where
 
-import System.IO (hPutStrLn)
-
 import Colog.Core.Action (LogAction (..))
 
 import qualified Data.ByteString.Char8 as BS
 import qualified Data.Text.IO as TIO
-
-----------------------------------------------------------------------------
--- String
-----------------------------------------------------------------------------
-
-{- | Action that prints 'String' to stdout. -}
-logStringStdout :: MonadIO m => LogAction m String
-logStringStdout = LogAction putStrLn
-
-{- | Action that prints 'String' to stderr. -}
-logStringStderr :: MonadIO m => LogAction m String
-logStringStderr = logStringHandle stderr
-
-{- | Action that prints 'String' to 'Handle'. -}
-logStringHandle :: MonadIO m => Handle -> LogAction m String
-logStringHandle handle = LogAction $ liftIO . hPutStrLn handle
-
-{- | Action that prints 'String' to file. Instead of returning 'LogAction' it's
-implemented in continuation-passing style because it's more efficient to open
-file only once at the start of the application and write to 'Handle' instead of
-opening file each time we need to write to it.
-
-Opens file in 'AppendMode'.
--}
-withLogStringFile :: MonadIO m => FilePath -> (LogAction m String -> IO r) -> IO r
-withLogStringFile path action = withFile path AppendMode $ action . logStringHandle
 
 ----------------------------------------------------------------------------
 -- ByteString